requirements.txt
py-cord
catapi.py
pygelbooru
pyspark